fix(editor): add `id` and set it as key
The new implementation of `EditorColor` that uses `vue3-colorpicker`
requires a `key` to be set. However, this key must not change since this
causes a re-render and color picker immediately disapears.

Composable `useQueryColors` now returns an array of objects containing
`color` -- the hex code, and `id` -- unique immutable value to be used
as key.